+++ plugins/destination/postgresql/client/pgx_log_level.go
@@ -1,4 +1,4 @@
-package postgresql
+package client
 
 import (
 	"bytes"